backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
openai==1.3.5
python-dotenv==1.0.0
boto3==1.29.3
celery==5.3.6
redis==5.0.1
flower==2.0.1
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
pdfplumber==0.10.3
